renaming to robot_mover_utils
