+++ droidlet/perception/robot/handlers/core.py
@@ -15,7 +15,7 @@ if "/opt/ros/kinetic/lib/python2.7/dist-packages" in sys.path:
 import cv2
 
 from droidlet.perception.robot.perception_helpers import get_color_tag
-from droidlet.lowlevel.locobot.locobot_mover_utils import xyz_pyrobot_to_canonical_coords
+from droidlet.lowlevel.robot_mover_utils import xyz_pyrobot_to_canonical_coords
 
 class AbstractHandler:
     """Interface for implementing perception handlers.
